Validation changes
Updates to validation to fit the comments from Scott. Validator Exceptions handled a bit better.
