+++ fastavro/_write_py.py
@@ -13,25 +13,22 @@ import decimal
 import time
 import uuid
 from binascii import crc32
-from collections import Iterable, Mapping
 from os import urandom, SEEK_SET
 from struct import pack
 from zlib import compress
-import numbers
 
-from fastavro._write_common import ValidationErrors, ValidationException
+from fastavro._validate_py import validate
 from .const import (
     MCS_PER_HOUR, MCS_PER_MINUTE, MCS_PER_SECOND, MLS_PER_HOUR, MLS_PER_MINUTE,
     MLS_PER_SECOND, DAYS_SHIFT
 )
-from .six import utob, MemoryIO, long, is_str, iterkeys, itervalues, \
-    iteritems, mk_bits
+from .six import utob, MemoryIO, long, iteritems, mk_bits
 from .read import HEADER_SCHEMA, SYNC_SIZE, MAGIC
 from .schema import (
     extract_named_schemas_into_repo, extract_record_type,
     extract_logical_type
 )
-from ._schema_common import SCHEMA_DEFS, UnknownType
+from ._schema_common import SCHEMA_DEFS
 from ._timezone import epoch
 
 NoneType = type(None)
@@ -311,174 +308,6 @@ INT_MIN_VALUE = -(1 << 31)
 INT_MAX_VALUE = (1 << 31) - 1
 LONG_MIN_VALUE = -(1 << 63)
 LONG_MAX_VALUE = (1 << 63) - 1
-
-
-def validate_null(datum, **kwargs):
-    return datum is None
-
-
-def validate_boolean(datum, **kwargs):
-    return isinstance(datum, bool)
-
-
-def validate_string(datum, **kwargs):
-    return is_str(datum)
-
-
-def validate_bytes(datum, **kwargs):
-    return isinstance(datum, (bytes, decimal.Decimal))
-
-
-def validate_int(datum, **kwargs):
-    return (
-        (isinstance(datum, (int, long, numbers.Integral)) and
-         INT_MIN_VALUE <= datum <= INT_MAX_VALUE) or
-        isinstance(datum, (datetime.time, datetime.datetime, datetime.date))
-    )
-
-
-def validate_long(datum, **kwargs):
-    return (
-        (isinstance(datum, (int, long, numbers.Integral)) and
-         LONG_MIN_VALUE <= datum <= LONG_MAX_VALUE) or
-        isinstance(datum, (datetime.time, datetime.datetime, datetime.date))
-    )
-
-
-def validate_float(datum, **kwargs):
-    return isinstance(datum, (int, long, float, numbers.Real))
-
-
-def validate_fixed(datum, schema, **kwargs):
-    return (
-            (isinstance(datum, bytes) and len(datum) == schema['size'])
-            or (isinstance(datum, decimal.Decimal))
-    )
-
-
-def validate_enum(datum, schema, **kwargs):
-    return datum in schema['symbols']
-
-
-def validate_array(datum, schema, raise_errors=False):
-    return (
-            isinstance(datum, Iterable) and
-            not is_str(datum) and
-            all(validate(datum=d, schema=schema['items'],
-                         field=schema.get('name'),
-                         raise_errors=raise_errors) for d in datum)
-    )
-
-
-def validate_map(datum, schema, raise_errors=False):
-    return (
-            isinstance(datum, Mapping) and
-            all(is_str(k) for k in iterkeys(datum)) and
-            all(validate(datum=v, schema=schema['values'],
-                         field=schema.get('name'),
-                         raise_errors=raise_errors) for v in itervalues(datum))
-    )
-
-
-def validate_record(datum, schema, raise_errors=False):
-    return (
-            isinstance(datum, Mapping) and
-            all(
-                validate(datum=datum.get(f['name'], f.get('default')),
-                         schema=f['type'],
-                         field=schema.get('name'),
-                         raise_errors=raise_errors)
-                for f in schema['fields']
-            )
-    )
-
-
-def validate_union(datum, schema, raise_errors=False):
-    if isinstance(datum, tuple):
-        (name, datum) = datum
-        for candidate in schema:
-            if extract_record_type(candidate) == 'record':
-                if name == candidate["name"]:
-                    return validate(datum, schema=candidate,
-                                    field=None,
-                                    raise_errors=raise_errors)
-        else:
-            return False
-
-    errors = []
-    for s in schema:
-        try:
-            ret = validate(datum, schema=s,
-                           field=None,
-                           raise_errors=raise_errors)
-            if ret:
-                # We exit on the first passing type in Unions
-                return True
-        except ValidationException as e:
-            errors.append(e.message)
-    if raise_errors:
-        raise ValidationErrors(errors)
-    return False
-
-
-VALIDATORS = {
-    'null': validate_null,
-    'boolean': validate_boolean,
-    'string': validate_string,
-    'int': validate_int,
-    'long': validate_long,
-    'float': validate_float,
-    'double': validate_float,
-    'bytes': validate_bytes,
-    'fixed': validate_fixed,
-    'enum': validate_enum,
-    'array': validate_array,
-    'map': validate_map,
-    'union': validate_union,
-    'error_union': validate_union,
-    'record': validate_record,
-    'error': validate_record,
-    'request': validate_record
-}
-
-
-def validate(datum, schema, field=None, raise_errors=False):
-    """Determine if a python datum is an instance of a schema."""
-    record_type = extract_record_type(schema)
-    result = None
-
-    if hasattr(schema, 'get'):
-        if field is not None:
-            ns_field = '.'.join([field, schema.get('name')])
-        else:
-            ns_field = schema.get('name')
-    else:
-        ns_field = field
-
-    if record_type in ('null', 'union'):
-        validator = VALIDATORS.get(record_type)
-        result = validator(datum, schema=schema, raise_errors=raise_errors)
-    elif datum is None:
-        # data required if not Null type
-        result = False
-    else:
-        validator = VALIDATORS.get(record_type)
-        if validator:
-            result = validator(datum, schema=schema, raise_errors=raise_errors)
-
-    if record_type in SCHEMA_DEFS and result is None:
-        result = validate(datum,
-                          schema=SCHEMA_DEFS[record_type],
-                          field=ns_field,
-                          raise_errors=raise_errors)
-
-    if raise_errors and result is False:
-        raise ValidationException(datum, schema, ns_field)
-
-    if result is None:
-        raise UnknownType(record_type)
-
-    return result
 
 
 def write_union(fo, datum, schema):
