+++ fastavro/schema.py
@@ -5,6 +5,7 @@ except ImportError:
 
 from ._schema_common import UnknownType
 
+schema_name = _schema.schema_name
 load_schema = _schema.load_schema
 extract_record_type = _schema.extract_record_type
 populate_schema_defs = _schema.populate_schema_defs
