+++ fastavro/validate.py
@@ -1,0 +1,10 @@
+try:
+    from . import _validate
+except ImportError:
+    from . import _validate_py as _validate
+from ._validate_common import ValidationError, ValidationErrors
+
+validate = _validate.validate
+VALIDATORS = _validate.VALIDATORS
+
+__all__ = ['ValidationErrors', 'ValidationError', 'validate', 'VALIDATORS']
