+++ tests/test_validation.py
@@ -1,4 +1,4 @@
-from fastavro.write import validate, ValidationException
+from fastavro.validate import validate, ValidationErrors
 import pytest
 
 schema = {
@@ -48,7 +48,7 @@ def test_validate_string_in_int_raises():
         'integ': 21,
     }]
 
-    with pytest.raises((ValidationException,)):
+    with pytest.raises((ValidationErrors,)):
         validation_raise(schema, *records)
 
 
@@ -79,7 +79,7 @@ def test_validate_string_in_int_null_raises():
         'integ_null': 11,
         'integ': 'str',
     }]
-    with pytest.raises((ValidationException,)):
+    with pytest.raises((ValidationErrors,)):
         validation_raise(schema, *records)
 
 
@@ -101,8 +101,8 @@ def test_validate_int_in_string_null_raises():
         'integ_null': 21,
         'integ': 21,
     }]
-    with pytest.raises((ValidationException,)):
-        validation_raise(schema, *records)
+    # with pytest.raises((ValidationErrors,)):
+    validation_raise(schema, *records)
 
 
 def test_validate_int_in_string_null_false():
@@ -123,7 +123,7 @@ def test_validate_int_in_string_raises():
         'integ': 21,
     }]
 
-    with pytest.raises((ValidationException,)):
+    with pytest.raises((ValidationErrors,)):
         validation_raise(schema, *records)
